+++ src/main/java/handler/request_handler/UserSignUpRequestHandler.java
@@ -1,4 +1,4 @@
-package handler;
+package handler.request_handler;
 
 import db.Database;
 import exception.*;
